examples/ai-stock-research-agent/requirements.txt
streamlit>=1.28.0
agno>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
bmasterai>=0.2.0
lxml>=4.9.0
html5lib>=1.1
firecrawl-py>=2.16.5